+++ smartsheet-csharp-sdk/main/Smartsheet/Api/Internal/AbstractResources.cs
@@ -186,9 +186,7 @@ namespace Smartsheet.Api.Internal
         /// <exception cref="SmartsheetException"> the SmartsheetClient exception </exception>
         protected internal virtual T GetResource<T>(string path, Type objectClass)
         {
-            var task = GetResourceAsync<T>(path, objectClass);
-            task.Wait();
-            return task.Result;
+            return GetResourceAsync<T>(path, objectClass).GetAwaiter().GetResult();
         }
 
         /// <summary>
@@ -287,9 +285,7 @@ namespace Smartsheet.Api.Internal
         /// <exception cref="SmartsheetException"> the SmartsheetClient exception </exception>
         protected internal virtual S CreateResource<S, T>(string path, T @object)
         {
-            var task = CreateResourceAsync<S, T>(path, @object);
-            task.Wait();
-            return task.Result;
+            return this.CreateResourceAsync<S, T>(path, @object).GetAwaiter().GetResult();
         }
 
         /// <summary>
@@ -362,9 +358,7 @@ namespace Smartsheet.Api.Internal
         /// <exception cref="SmartsheetException"> the SmartsheetClient exception </exception>
         protected internal virtual T CreateResource<T>(string path, Type objectClass, T @object)
         {
-            var task = CreateResourceAsync<T>(path, objectClass, @object);
-            task.Wait();
-            return task.Result;
+            return this.CreateResourceAsync<T>(path, objectClass, @object).GetAwaiter().GetResult();
         }
 
         /// <summary>
@@ -432,6 +426,22 @@ namespace Smartsheet.Api.Internal
         /// <exception cref="InvalidOperationException"> if any argument is null, or path is an empty string </exception>
         protected internal virtual T CreateResourceWithAttachment<T>(string path, T @object, string objectType, string file, string fileType)
         {
+            return this.CreateResourceWithAttachmentAsync(path, @object, objectType, file, fileType).GetAwaiter().GetResult();
+        }
+
+        /// <summary>
+        /// Create a resource with an attachment.
+        /// </summary>
+        /// <param name="path"> the relative path of the resource collections </param>
+        /// <param name="object">the object to create </param>
+        /// <param name="file"> the file path </param>
+        /// <param name="fileType"> the file type, can be null </param>
+        /// <param name="cancellationToken"> the cancellation token </param>
+        /// <param name="objectType"> the object type to return as </param>
+        /// <returns> the created resource </returns>
+        /// <exception cref="InvalidOperationException"> if any argument is null, or path is an empty string </exception>
+        protected async internal virtual Task<T> CreateResourceWithAttachmentAsync<T>(string path, T @object, string objectType, string file, string fileType, CancellationToken cancellationToken = default)
+        {
             Utils.ThrowIfNull(path, @object);
             Utils.ThrowIfEmpty(path);
 
@@ -446,7 +456,7 @@ namespace Smartsheet.Api.Internal
             }
 
             request.Entity = serializeToEntity<T>(@object);
-            HttpResponse response = this.smartsheet.HttpClient.Request(request, objectType, file, fileType);
+            HttpResponse response = await this.smartsheet.HttpClient.RequestAsync(request, objectType, file, fileType, cancellationToken).ConfigureAwait(false);
 
             Object obj = null;
             switch (response.StatusCode)
@@ -484,9 +494,7 @@ namespace Smartsheet.Api.Internal
         /// <exception cref="SmartsheetException"> the SmartsheetClient exception </exception>
         protected internal virtual T UpdateResource<T>(string path, Type objectClass, T @object)
         {
-            var task = UpdateResourceAsync<T>(path, objectClass, @object);
-            task.Wait();
-            return task.Result;
+            return this.UpdateResourceAsync<T>(path, objectClass, @object).GetAwaiter().GetResult();
         }
 
         /// <summary>
@@ -609,6 +617,26 @@ namespace Smartsheet.Api.Internal
         /// <exception cref="SmartsheetException"> if an error occurred during the operation </exception>
         protected internal virtual PaginatedResult<T> ListResourcesWithWrapper<T>(string path)
         {
+            return this.ListResourcesWithWrapperAsync<T>(path).GetAwaiter().GetResult();
+        }
+
+        /// <summary>
+        /// Asynchronously list resources using SmartsheetClient REST API.
+        /// 
+        /// Exceptions:
+        ///   IllegalArgumentException : if any argument is null, or path is an empty string
+        ///   InvalidRequestException : if there is any problem with the REST API request
+        ///   AuthorizationException : if there is any problem with the REST API authorization (access token)
+        ///   ServiceUnavailableException : if the REST API service is not available (possibly due to rate limiting)
+        ///   SmartsheetRestException : if any other REST API related error occurred during the operation
+        ///   SmartsheetException : if any other error occurred during the operation
+        /// </summary>
+        /// <param name="path"> the relative path of the resource collections </param>
+        /// <param name="cancellationToken"> the cancellation token </param>
+        /// <returns> the resources </returns>
+        /// <exception cref="SmartsheetException"> if an error occurred during the operation </exception>
+        protected internal async virtual Task<PaginatedResult<T>> ListResourcesWithWrapperAsync<T>(string path, CancellationToken cancellationToken = default)
+        {
             Utils.ThrowIfNull(path);
             Utils.ThrowIfEmpty(path);
 
@@ -622,7 +650,7 @@ namespace Smartsheet.Api.Internal
                 throw new SmartsheetException(e);
             }
 
-            HttpResponse response = this.smartsheet.HttpClient.Request(request);
+            HttpResponse response = await this.smartsheet.HttpClient.RequestAsync(request, cancellationToken).ConfigureAwait(false);
 
             PaginatedResult<T> obj = null;
             switch (response.StatusCode)
@@ -657,6 +685,26 @@ namespace Smartsheet.Api.Internal
         /// <exception cref="SmartsheetException"> if an error occurred during the operation </exception>
         protected internal virtual TokenPaginatedResult<T> ListResourcesWithTokenWrapper<T>(string path)
         {
+            return this.ListResourcesWithTokenWrapperAsync<T>(path).GetAwaiter().GetResult();
+        }
+
+        /// <summary>
+        /// List resources using SmartsheetClient REST API with token-based pagination.
+        /// 
+        /// Exceptions:
+        ///   IllegalArgumentException : if any argument is null, or path is an empty string
+        ///   InvalidRequestException : if there is any problem with the REST API request
+        ///   AuthorizationException : if there is any problem with the REST API authorization (access token)
+        ///   ServiceUnavailableException : if the REST API service is not available (possibly due to rate limiting)
+        ///   SmartsheetRestException : if any other REST API related error occurred during the operation
+        ///   SmartsheetException : if any other error occurred during the operation
+        /// </summary>
+        /// <param name="path"> the relative path of the resource collections </param>
+        /// <param name="cancellationToken"> the cancellation token </param>
+        /// <returns> the resources with token-based pagination </returns>
+        /// <exception cref="SmartsheetException"> if an error occurred during the operation </exception>
+        protected internal async virtual Task<TokenPaginatedResult<T>> ListResourcesWithTokenWrapperAsync<T>(string path, CancellationToken cancellationToken = default)
+        {
             Utils.ThrowIfNull(path);
             Utils.ThrowIfEmpty(path);
 
@@ -670,7 +718,7 @@ namespace Smartsheet.Api.Internal
                 throw new SmartsheetException(e);
             }
 
-            HttpResponse response = this.smartsheet.HttpClient.Request(request);
+            HttpResponse response = await this.smartsheet.HttpClient.RequestAsync(request, cancellationToken).ConfigureAwait(false);
 
             TokenPaginatedResult<T> obj = null;
             switch (response.StatusCode)
@@ -705,6 +753,26 @@ namespace Smartsheet.Api.Internal
         /// <exception cref="SmartsheetException"> if an error occurred during the operation </exception>
         protected internal virtual IList<T> ListResources<T>(string path, Type objectClass)
         {
+            return this.ListResourcesAsync<T>(path, objectClass).GetAwaiter().GetResult();
+        }
+
+        /// <summary>
+        /// List resources using SmartsheetClient REST API.
+        /// 
+        /// Exceptions:
+        ///   IllegalArgumentException : if any argument is null, or path is an empty string
+        ///   InvalidRequestException : if there is any problem with the REST API request
+        ///   AuthorizationException : if there is any problem with the REST API authorization (access token)
+        ///   ServiceUnavailableException : if the REST API service is not available (possibly due to rate limiting)
+        ///   SmartsheetRestException : if any other REST API related error occurred during the operation
+        ///   SmartsheetException : if any other error occurred during the operation
+        /// </summary>
+        /// <param name="path"> the relative path of the resource collections </param>
+        /// <param name="objectClass"> the resource object class </param>
+        /// <returns> the resources </returns>
+        /// <exception cref="SmartsheetException"> if an error occurred during the operation </exception>
+        protected async internal virtual Task<IList<T>> ListResourcesAsync<T>(string path, Type objectClass, CancellationToken cancellationToken = default)
+        {
             Utils.ThrowIfNull(path, objectClass);
             Utils.ThrowIfEmpty(path);
 
@@ -718,7 +786,7 @@ namespace Smartsheet.Api.Internal
                 throw new SmartsheetException(e);
             }
 
-            HttpResponse response = this.smartsheet.HttpClient.Request(request);
+            HttpResponse response = await this.smartsheet.HttpClient.RequestAsync(request, cancellationToken).ConfigureAwait(false);
 
             IList<T> obj = null;
             switch (response.StatusCode)
@@ -753,6 +821,26 @@ namespace Smartsheet.Api.Internal
         /// <exception cref="SmartsheetException"> the SmartsheetClient exception </exception>
         protected internal virtual T DeleteResource<T>(string path)
         {
+            return this.DeleteResourceAsync<T>(path).GetAwaiter().GetResult();
+        }
+
+        /// <summary>
+        /// Delete a resource from SmartsheetClient REST API.
+        /// 
+        /// Exceptions:
+        ///   IllegalArgumentException : if any argument is null, or path is an empty string
+        ///   InvalidRequestException : if there is any problem with the REST API request
+        ///   AuthorizationException : if there is any problem with the REST API authorization (access token)
+        ///   ResourceNotFoundException : if the resource cannot be found
+        ///   ServiceUnavailableException : if the REST API service is not available (possibly due to rate limiting)
+        ///   SmartsheetRestException : if any other REST API related error occurred during the operation
+        ///   SmartsheetException : if any other error occurred during the operation
+        /// </summary>
+        /// <param name="path"> the relative path of the resource </param>
+        /// <param name="cancellationToken"> the cancellation token </param>
+        /// <exception cref="SmartsheetException"> the SmartsheetClient exception </exception>
+        protected async internal virtual Task<T> DeleteResourceAsync<T>(string path, CancellationToken cancellationToken = default)
+        {
             Utils.ThrowIfNull(path);
             Utils.ThrowIfEmpty(path);
 
@@ -765,7 +853,7 @@ namespace Smartsheet.Api.Internal
             {
                 throw new SmartsheetException(e);
             }
-            HttpResponse response = this.smartsheet.HttpClient.Request(request);
+            HttpResponse response = await this.smartsheet.HttpClient.RequestAsync(request, cancellationToken).ConfigureAwait(false);
             Object obj = null;
             switch (response.StatusCode)
             {
@@ -799,6 +887,27 @@ namespace Smartsheet.Api.Internal
         /// <exception cref="SmartsheetException"> the SmartsheetClient exception </exception>
         protected internal virtual void DeleteResource<T>(string path, Type objectClass)
         {
+            this.DeleteResourceAsync<T>(path, objectClass).Wait();
+        }
+
+        /// <summary>
+        /// Asynchronously delete a resource from SmartsheetClient REST API.
+        /// 
+        /// Exceptions:
+        ///   IllegalArgumentException : if any argument is null, or path is an empty string
+        ///   InvalidRequestException : if there is any problem with the REST API request
+        ///   AuthorizationException : if there is any problem with the REST API authorization (access token)
+        ///   ResourceNotFoundException : if the resource cannot be found
+        ///   ServiceUnavailableException : if the REST API service is not available (possibly due to rate limiting)
+        ///   SmartsheetRestException : if any other REST API related error occurred during the operation
+        ///   SmartsheetException : if any other error occurred during the operation
+        /// </summary>
+        /// <param name="path"> the relative path of the resource </param>
+        /// <param name="objectClass"> the resource object class </param>
+        /// <param name="cancellationToken"> the cancellation token </param>
+        /// <exception cref="SmartsheetException"> the SmartsheetClient exception </exception>
+        protected internal async virtual Task DeleteResourceAsync<T>(string path, Type objectClass, CancellationToken cancellationToken = default)
+        {
             Utils.ThrowIfNull(path, objectClass);
             Utils.ThrowIfEmpty(path);
 
@@ -811,7 +920,7 @@ namespace Smartsheet.Api.Internal
             {
                 throw new SmartsheetException(e);
             }
-            HttpResponse response = this.smartsheet.HttpClient.Request(request);
+            HttpResponse response = await this.smartsheet.HttpClient.RequestAsync(request, cancellationToken).ConfigureAwait(false);
 
             switch (response.StatusCode)
             {